Move account routes to proper directories
